requirements.txt
# Dependencies  
django
djangorestframework
transformers
torch
soundfile
torchaudio
protobuf
sentencepiece
opencv-python
pillow